+++ tests/local_onionshare_receive_mode_upload_public_mode_test.py
@@ -18,6 +18,7 @@ class LocalReceiveModePublicModeTest(unittest.TestCase, GuiReceiveTest):
         GuiReceiveTest.tear_down()
 
     @pytest.mark.gui
+    @pytest.mark.skipif(pytest.__version__ < '2.9', reason="requires newer pytest")
     def test_gui(self):
         self.run_all_common_setup_tests()
         self.run_all_receive_mode_tests(True, True)