+++ tests/local_onionshare_receive_mode_upload_test.py
@@ -17,6 +17,7 @@ class LocalReceiveModeTest(unittest.TestCase, GuiReceiveTest):
         GuiReceiveTest.tear_down()
 
     @pytest.mark.gui
+    @pytest.mark.skipif(pytest.__version__ < '2.9', reason="requires newer pytest")
     def test_gui(self):
         self.run_all_common_setup_tests()
         self.run_all_receive_mode_tests(False, True)